connecting with back

+++ src/api/basics/users/deleteUser.js
@@ -1,7 +1,7 @@
 import axiosClient from "../../../libs/axios/axios-client";
 
 async function deleteUsers({ userId }) {
-  const res = await axiosClient.delete();
+  const res = await axiosClient.delete(`/users/${userId}`);
 
   return res?.data;
 }
